修改：整理

+++ src/main/java/org/rock/base/common/mongo/convertConfig/BigDecimalToDecimal128Converter.java
@@ -17,7 +17,7 @@ public class BigDecimalToDecimal128Converter implements Converter<BigDecimal, De
 
     @Override
     public Decimal128 convert(BigDecimal bigDecimal) {
-        return new Decimal128(BigDecimal.valueOf(bigDecimal.doubleValue()));
+        return new Decimal128(bigDecimal);
     }
 
 }
